+++ jgclark.WindowTools/src/openers.js
@@ -2,15 +2,16 @@
 //---------------------------------------------------------------
 // Various open window/split functions for WindowTools plugin
 // Jonathan Clark
-// last update 2025-08-03 for v1.2.2 by @jgclark
+// last update 2025-11-30 for v1.5.0 by @jgclark
 //---------------------------------------------------------------
 
 import pluginJson from '../plugin.json'
+import { getPluginSettings } from './WTHelpers'
 import { clo, JSP, logDebug, logError, logInfo, logWarn } from '@helpers/dev'
-// import { allNotesSortedByChanged } from '@helpers/note'
-import { getNoteFromIdentifier } from '@helpers/NPnote'
+import { chooseNoteV2, getNoteFromIdentifier } from '@helpers/NPnote'
 import { findStartOfActivePartOfNote } from '@helpers/paragraph'
-import { chooseNoteV2 } from '@helpers/NPnote'
+import { openNoteInNewWindow } from '@helpers/NPWindows'
+import { showMessage } from '@helpers/userInput'
 
 /**
  * Open a user-selected note in a new window.
@@ -21,6 +22,10 @@ import { chooseNoteV2 } from '@helpers/NPnote'
  */
 export async function openNoteNewWindow(encodedNoteIdentifier: string = ''): Promise<void> {
   try {
+    if (NotePlan.environment.platform !== 'macOS') {
+      await showMessage(`This command can only run on macOS. Stopping.`, 'OK', 'Error')
+      return
+    }
     let note: ?TNote
     if (encodedNoteIdentifier !== '') {
       note = getNoteFromIdentifier(decodeURIComponent(encodedNoteIdentifier))
@@ -31,13 +36,24 @@ export async function openNoteNewWindow(encodedNoteIdentifier: string = ''): Pro
     }
     if (note) {
       const filename = note.filename
+      const config = await getPluginSettings()
+
+      // Up to v1.2.x ...
       // work out where start of main content of the note is
-      const startOfMainContentLine = findStartOfActivePartOfNote(note)
-      const startOfMainContentCharIndex = (startOfMainContentLine > 0)
-        ? note.paragraphs[startOfMainContentLine].contentRange?.start ?? 0
-        : 0
+      // const startOfMainContentLine = findStartOfActivePartOfNote(note)
+      // const startOfMainContentCharIndex = (startOfMainContentLine > 0)
+      //   ? note.paragraphs[startOfMainContentLine].contentRange?.start ?? 0
+      //   : 0
       // open note, moving cursor to start of main content
-      const res = await Editor.openNoteByFilename(filename, true, startOfMainContentCharIndex, startOfMainContentCharIndex, false)
+      // const res = await Editor.openNoteByFilename(filename, true, startOfMainContentCharIndex, startOfMainContentCharIndex, false)
+
+      // From v1.5.0 ...
+      // open note, using smart features to place the window on the screen
+      if (config.useSmartPlacement) {
+        const res = await openNoteInNewWindow(filename, config.defaultEditorWidth ?? 0, false, true)
+      } else {
+        const res = await openNoteInNewWindow(filename, config.defaultEditorWidth ?? 0, true, false)
+      }
     } else {
       logWarn(pluginJson, `openNoteNewWindow: Couldn't find note to open`)
     }
@@ -55,6 +71,10 @@ export async function openNoteNewWindow(encodedNoteIdentifier: string = ''): Pro
  */
 export async function openNoteNewSplit(encodedNoteIdentifier: string = ''): Promise<void> {
   try {
+    if (NotePlan.environment.platform !== 'macOS') {
+      await showMessage(`This command can only run on macOS. Stopping.`, 'OK', 'Error')
+      return
+    }
     let note: ?TNote
     if (encodedNoteIdentifier !== '') {
       note = getNoteFromIdentifier(decodeURIComponent(encodedNoteIdentifier))
@@ -86,6 +106,10 @@ export async function openNoteNewSplit(encodedNoteIdentifier: string = ''): Prom
  */
 export async function openCurrentNoteNewSplit(): Promise<void> {
   try {
+    if (NotePlan.environment.platform !== 'macOS') {
+      await showMessage(`This command can only run on macOS. Stopping.`, 'OK', 'Error')
+      return
+    }
     const { note, filename } = Editor
     if (note == null || filename == null) {
       // No note open, so don't do anything.
@@ -108,19 +132,32 @@ export async function openCurrentNoteNewSplit(): Promise<void> {
  */
 export async function openCurrentNoteNewWindow(): Promise<void> {
   try {
+    if (NotePlan.environment.platform !== 'macOS') {
+      await showMessage(`This command can only run on macOS. Stopping.`, 'OK', 'Error')
+      return
+    }
     const { note, filename } = Editor
     if (note == null || filename == null) {
       // No note open, so don't do anything.
       logError('openCurrentNoteNewSplit()', 'No note open. Stopping.')
       return
     }
+    const config = await getPluginSettings()
+    // Up to v1.2.x ...
     // work out where start of main content of the note is
-    const startOfMainContentLine = findStartOfActivePartOfNote(note)
-    const startOfMainContentCharIndex = note.paragraphs[startOfMainContentLine].contentRange?.start ?? 0
+    // const startOfMainContentLine = findStartOfActivePartOfNote(note)
+    // const startOfMainContentCharIndex = note.paragraphs[startOfMainContentLine].contentRange?.start ?? 0
     // open note, moving cursor to start of main content
-    await Editor.openNoteByFilename(filename, true, startOfMainContentCharIndex, startOfMainContentCharIndex, false)
-  } catch (e) {
-    logError('openCurrentNoteNewSplit()', e.message)
+    // await Editor.openNoteByFilename(filename, true, startOfMainContentCharIndex, startOfMainContentCharIndex, false)
+
+    // From v1.5.0 ...
+    // open note, using smart features to place the window on the screen
+    if (config.useSmartPlacement) {
+      const res = await openNoteInNewWindow(filename, config.defaultEditorWidth ?? 0, false, true)
+    } else {
+      const res = await openNoteInNewWindow(filename, config.defaultEditorWidth ?? 0, true, false)
+    }
+  } catch (err) {
+    logError('openCurrentNoteNewWindow()', err.message)
   }
 }
-